final version

+++ SpringProject/src/main/java/com/example/spring_test/repository/CollectionRepository.java
@@ -3,5 +3,5 @@ package com.example.spring_test.repository;
 import com.example.spring_test.models.Collection;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CollectionRepository extends JpaRepository<Collection, Integer> {
+public interface CollectionRepository extends JpaRepository<Collection, Long> {
 }
