+++ SpringProject/src/main/java/com/example/spring_test/repository/DonationRepository.java
@@ -3,5 +3,5 @@ package com.example.spring_test.repository;
 import com.example.spring_test.models.Donation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DonationRepository extends JpaRepository<Donation, Integer> {
+public interface DonationRepository extends JpaRepository<Donation, Long> {
 }
